+++ x/marker/handler_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 
 	"github.com/provenance-io/provenance/x/marker"
 	"github.com/provenance-io/provenance/x/marker/keeper"
@@ -24,4 +25,13 @@ func TestInvalidMsg(t *testing.T) {
 	require.Error(t, err)
 	require.Nil(t, res)
 	require.True(t, strings.Contains(err.Error(), "unknown message type: Test message"))
+}
+
+func TestInvalidProposal(t *testing.T) {
+	k := keeper.Keeper{}
+	h := marker.NewProposalHandler(k)
+
+	err := h(sdk.NewContext(nil, tmproto.Header{}, false, nil), govtypes.NewTextProposal("Test", "description"))
+	require.Error(t, err)
+	require.True(t, strings.Contains(err.Error(), "unrecognized marker proposal content type: *types.TextProposal"))
 }